requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
transformers
torch
pydantic
numpy
google-generativeai
python-dotenv